+++ src/game/g_trigger.cpp
@@ -269,6 +269,18 @@ trigger_push
 
 void trigger_push_touch(gentity_t *self, gentity_t *other, trace_t *trace)
 {
+	// Disable for specs
+	if (!other->client)
+	{
+		return;
+	}
+
+	if (self->noise_index)
+	{
+		G_AddEvent(self, EV_GENERAL_SOUND, self->noise_index);
+	}
+	
+	BG_TouchJumpPad(&other->client->ps, &self->s);
 }
 
 
@@ -326,6 +338,22 @@ This will be client side predicted, unlike target_push
 */
 void SP_trigger_push(gentity_t *self)
 {
+	char *s;
+
+	InitTrigger(self);
+
+	// Send to client for prediction
+	self->r.svFlags &= ~SVF_NOCLIENT;
+
+	// Noise key support
+	G_SpawnString("noise", "", &s);
+	self->noise_index = G_SoundIndex(s);
+
+	self->s.eType = ET_PUSH_TRIGGER;
+	self->touch = trigger_push_touch;
+	self->think = AimAtTarget;
+	self->nextthink = level.time + FRAMETIME;
+	trap_LinkEntity(self);
 }